renamed subscription dto vaidator and updated tests
