+++ test/integration/weather.test.js
@@ -3,7 +3,7 @@ const request = require('supertest');
 const express = require('express');
 
 const WeatherApiController = require('../../src/presentation/controllers/weatherApiController');
-const WeatherServiceWithCacheAndMetrics = require('../../src/services/weatherService');
+const WeatherServiceWithCacheAndMetrics = require('../../src/application/services/weatherService');
 const GetWeatherUseCase = require('../../src/application/use-cases/weather/getWeatherUseCase');
 const MockWeatherProviderManager = require('../mocks/providers/weatherProviderManager.mock');
 const { redisClient, connectToRedisWithRetry } = require('../../src/common/utils/redisClient');
