+++ test/unit/validators/subscriptionValidator.test.js
@@ -1,7 +1,7 @@
 const { expect } = require('chai');
-const SubscriptionValidator = require('../../../src/presentation/validators/subscriptionValidator');
+const SubscriptionDtoValidator = require('../../../src/presentation/validators/subscriptionDtoValidator');
 
-const subscriptionValidator = new SubscriptionValidator();
+const subscriptionDtoValidator = new SubscriptionDtoValidator();
 
 const validSub = {
     email: 'valid@mail.com',
@@ -9,12 +9,12 @@ const validSub = {
     frequency: 'hourly'
 };
 
-describe('SubscriptionValidator Unit Tests', () => 
+describe('subscriptionDtoValidator Unit Tests', () =>
 {
     it('should return true for success for valid subscription data', async () => 
     {
         // Act
-        const result = await subscriptionValidator.validateNewSubscription(validSub.email, validSub.city, validSub.frequency);
+        const result = subscriptionDtoValidator.validateNewSubscription(validSub.email, validSub.city, validSub.frequency);
 
         // Assert
         expect(result.success).to.be.true;
@@ -23,7 +23,7 @@ describe('SubscriptionValidator Unit Tests', () =>
     it('should return false for success and an error message for missing email', async () => 
     {
         // Act
-        const result = await subscriptionValidator.validateNewSubscription(null, validSub.city, validSub.frequency);
+        const result = subscriptionDtoValidator.validateNewSubscription(null, validSub.city, validSub.frequency);
 
         // Assert
         expect(result.success).to.be.false;
@@ -33,7 +33,7 @@ describe('SubscriptionValidator Unit Tests', () =>
     it('should return false for success and an error message for missing city', async () => 
     {
         // Act
-        const result = await subscriptionValidator.validateNewSubscription(validSub.email, null, validSub.frequency);
+        const result = subscriptionDtoValidator.validateNewSubscription(validSub.email, null, validSub.frequency);
 
         // Assert
         expect(result.success).to.be.false;
@@ -43,7 +43,7 @@ describe('SubscriptionValidator Unit Tests', () =>
     it('should return false for success and an error message for missing frequency', async () => 
     {
         // Act
-        const result = await subscriptionValidator.validateNewSubscription(validSub.email, validSub.city, null);
+        const result = subscriptionDtoValidator.validateNewSubscription(validSub.email, validSub.city, null);
 
         // Assert
         expect(result.success).to.be.false;
@@ -53,7 +53,7 @@ describe('SubscriptionValidator Unit Tests', () =>
     it('should return false for success and an error message for invalid email format', async () => 
     {
         // Act
-        const result = await subscriptionValidator.validateNewSubscription('invalid-email', validSub.city, validSub.frequency);
+        const result = subscriptionDtoValidator.validateNewSubscription('invalid-email', validSub.city, validSub.frequency);
 
         // Assert
         expect(result.success).to.be.false;
@@ -63,7 +63,7 @@ describe('SubscriptionValidator Unit Tests', () =>
     it('should return false for success and an error message for invalid frequency', async () => 
     {
         // Act
-        const result = await subscriptionValidator.validateNewSubscription(validSub.email, validSub.city, 'apple');
+        const result = subscriptionDtoValidator.validateNewSubscription(validSub.email, validSub.city, 'apple');
 
         // Assert
         expect(result.success).to.be.false;
